libebml: EbmlElement::VoidMe() now returns a uint64
git-svn-id: https://matroska.svn.sourceforge.net/svnroot/matroska/trunk/libebml@364 a6f86f6d-0131-4f8e-9e7b-e335508773d5

+++ ebml/EbmlElement.h
@@ -442,7 +442,7 @@ class EBML_DLL_API EbmlElement {
 		/*!
 			\brief void the content of the element (replace by EbmlVoid)
 		*/
-		uint32 VoidMe(IOCallback & output, bool bWithDefault = false);
+		uint64 VoidMe(IOCallback & output, bool bWithDefault = false);
 
 		bool DefaultISset() const {return DefaultIsSet;}
 		virtual bool IsDefaultValue() const = 0;
